# molenapplicatie.client/swagger/swagger.backup.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MolenApplicatie.Server",
    "version": "1.0"
  },
  "paths": {
    "/api/DB/all": {
      "get": {
        "tags": ["DB"],
        "operationId": "DB_GetAllDataForDB_GET_api_DB_all",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "DB",
          "actionName": "GetAllDataForDB"
        }
      }
    },
    "/api/DB/places": {
      "get": {
        "tags": ["DB"],
        "operationId": "DB_GetPlacesDB_GET_api_DB_places",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "DB",
          "actionName": "GetPlacesDB"
        }
      }
    },
    "/api/DB/GetMolens": {
      "get": {
        "tags": ["DB"],
        "operationId": "DB_GetMolens_GET_api_DB_GetMolens",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MolenData"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MolenData"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MolenData"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "DB",
          "actionName": "GetMolens"
        }
      }
    },
    "/api/DB/CallMolenResponses": {
      "get": {
        "tags": ["DB"],
        "operationId": "DB_CallMolenResponses_GET_api_DB_CallMolenResponses",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "DB",
          "actionName": "CallMolenResponses"
        }
      }
    },
    "/api/DB/SaveMolenResponses": {
      "get": {
        "tags": ["DB"],
        "operationId": "DB_SaveMolenResponses_GET_api_DB_SaveMolenResponses",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "DB",
          "actionName": "SaveMolenResponses"
        }
      }
    },
    "/api/DB/test": {
      "get": {
        "tags": ["DB"],
        "operationId": "DB_test_GET_api_DB_test",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "DB",
          "actionName": "test"
        }
      }
    },
    "/api/DB/test2": {
      "get": {
        "tags": ["DB"],
        "operationId": "DB_test2_GET_api_DB_test2",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "DB",
          "actionName": "test2"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddedImage": {
        "required": [
          "canBeDeleted",
          "filePath",
          "id",
          "molenData",
          "molenDataId",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "filePath": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "canBeDeleted": {
            "type": "boolean"
          },
          "dateTaken": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "molenData": {
            "$ref": "#/components/schemas/MolenData"
          },
          "molenDataId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "DisappearedYearInfo": {
        "required": ["id", "molenData", "molenDataId", "year"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status_before": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "status_after": {
            "type": "string",
            "nullable": true
          },
          "molenData": {
            "$ref": "#/components/schemas/MolenData"
          },
          "molenDataId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "HttpValidationProblemDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          },
          {
            "required": ["errors"],
            "type": "object",
            "properties": {
              "errors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "additionalProperties": {}
          }
        ]
      },
...
      "MolenData": {
        "required": [
          "addedImages",
          "canAddImages",
          "disappearedYearInfos",
          "hasImage",
          "id",
          "images",
          "lastUpdated",
          "latitude",
          "longitude",
          "mercatorY",
          "molenMakers",
          "molenTBN",
          "molenTBNId",
          "molenTypeAssociations",
          "name",
          "openOpAfspraak",
          "openOpZaterdag",
          "openOpZondag",
          "openVoorPubliek",
          "ten_Brugge_Nr"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "molenTBN": {
            "$ref": "#/components/schemas/MolenTBN"
          },
          "molenTBNId": {
            "type": "string",
            "format": "uuid"
          },
          "ten_Brugge_Nr": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "toelichtingNaam": {
            "type": "string",
            "nullable": true
          },
          "bouwjaar": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "herbouwdJaar": {
            "type": "string",
            "nullable": true
          },
          "bouwjaarStart": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bouwjaarEinde": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "functie": {
            "type": "string",
            "nullable": true
          },
          "doel": {
            "type": "string",
            "nullable": true
          },
          "toestand": {
            "type": "string",
            "nullable": true
          },
          "bedrijfsvaardigheid": {
            "type": "string",
            "nullable": true
          },
          "plaats": {
            "type": "string",
            "nullable": true
          },
          "adres": {
            "type": "string",
            "nullable": true
          },
          "land": {
            "type": "string",
            "nullable": true
          },
          "provincie": {
            "type": "string",
            "nullable": true
          },
          "gemeente": {
            "type": "string",
            "nullable": true
          },
          "streek": {
            "type": "string",
            "nullable": true
          },
          "plaatsaanduiding": {
            "type": "string",
            "nullable": true
          },
          "opvolger": {
            "type": "string",
            "nullable": true
          },
          "voorganger": {
            "type": "string",
            "nullable": true
          },
          "opvolgerMolenId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "voorgangerMolenId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "verplaatstNaar": {
            "type": "string",
            "nullable": true
          },
          "afkomstigVan": {
            "type": "string",
            "nullable": true
          },
          "literatuur": {
            "type": "string",
            "nullable": true
          },
          "plaatsenVoorheen": {
            "type": "string",
            "nullable": true
          },
          "wiekvorm": {
            "type": "string",
            "nullable": true
          },
          "wiekVerbeteringen": {
            "type": "string",
            "nullable": true
          },
          "monument": {
            "type": "string",
            "nullable": true
          },
          "plaatsBediening": {
            "type": "string",
            "nullable": true
          },
          "bedieningKruiwerk": {
            "type": "string",
            "nullable": true
          },
          "plaatsKruiwerk": {
            "type": "string",
            "nullable": true
          },
          "kruiwerk": {
            "type": "string",
            "nullable": true
          },
          "vlucht": {
            "type": "string",
            "nullable": true
          },
          "openingstijden": {
            "type": "string",
            "nullable": true
          },
          "openVoorPubliek": {
            "type": "boolean"
          },
          "openOpZaterdag": {
            "type": "boolean"
          },
          "openOpZondag": {
            "type": "boolean"
          },
          "openOpAfspraak": {
            "type": "boolean"
          },
          "krachtbron": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "winkelInformatie": {
            "type": "string",
            "nullable": true
          },
          "bouwbestek": {
            "type": "string",
            "nullable": true
          },
          "bijzonderheden": {
            "type": "string",
            "nullable": true
          },
          "museuminformatie": {
            "type": "string",
            "nullable": true
          },
          "molenaar": {
            "type": "string",
            "nullable": true
          },
          "eigendomshistorie": {
            "type": "string",
            "nullable": true
          },
          "molenerf": {
            "type": "string",
            "nullable": true
          },
          "trivia": {
            "type": "string",
            "nullable": true
          },
          "geschiedenis": {
            "type": "string",
            "nullable": true
          },
          "wetenswaardigheden": {
            "type": "string",
            "nullable": true
          },
          "wederopbouw": {
            "type": "string",
            "nullable": true
          },
          "as": {
            "type": "string",
            "nullable": true
          },
          "wieken": {
            "type": "string",
            "nullable": true
          },
          "toegangsprijzen": {
            "type": "string",
            "nullable": true
          },
          "uniekeEigenschap": {
            "type": "string",
            "nullable": true
          },
          "landschappelijkeWaarde": {
            "type": "string",
            "nullable": true
          },
          "kadastraleAanduiding": {
            "type": "string",
            "nullable": true
          },
          "canAddImages": {
            "type": "boolean"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MolenImage"
            }
          },
          "molenTypeAssociations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MolenTypeAssociation"
            }
          },
          "addedImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddedImage"
            }
          },
          "molenMakers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MolenMaker"
            }
          },
          "disappearedYearInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisappearedYearInfo"
            }
          },
          "hasImage": {
            "type": "boolean"
          },
          "eigenaar": {
            "type": "string",
            "nullable": true
          },
          "recenteWerkzaamheden": {
            "type": "string",
            "nullable": true
          },
          "rad": {
            "type": "string",
            "nullable": true
          },
          "radDiameter": {
            "type": "string",
            "nullable": true
          },
          "wateras": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "mercatorY": {
            "type": "number",
            "format": "double"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "MolenImage": {
        "required": [
          "canBeDeleted",
          "externalUrl",
          "filePath",
          "id",
          "molenData",
          "molenDataId",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "filePath": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "canBeDeleted": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "externalUrl": {
            "type": "string"
          },
          "molenData": {
            "$ref": "#/components/schemas/MolenData"
          },
          "molenDataId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "MolenMaker": {
        "required": ["id", "molenData", "molenDataId", "name", "year"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "year": {
            "type": "string"
          },
          "molenData": {
            "$ref": "#/components/schemas/MolenData"
          },
          "molenDataId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "MolenTBN": {
        "required": ["id", "molenData", "ten_Brugge_Nr"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ten_Brugge_Nr": {
            "type": "string"
          },
          "molenData": {
            "$ref": "#/components/schemas/MolenData"
          }
        },
        "additionalProperties": false
      },
      "MolenType": {
        "required": ["id", "name"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "MolenTypeAssociation": {
        "required": [
          "id",
          "molenData",
          "molenDataId",
          "molenType",
          "molenTypeId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "molenDataId": {
            "type": "string",
            "format": "uuid"
          },
          "molenTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "molenData": {
            "$ref": "#/components/schemas/MolenData"
          },
          "molenType": {
            "$ref": "#/components/schemas/MolenType"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
    }
  },
...
}